Correct bug polarity and add utility for sorting polarities
For calculating piercing points and predicting polarities MTUQ used to assume that 'p' wave arrives first that 'P' when both phases are present, which is not true and cause miscalculation of piercing points and predicted polarities when 'P' wave arrives first than 'p' wave.

I also added an utility for sorting the polarity array according to the order in data list, after reading the data.